Update registry name in makefile
When installing a development version, the plugin currently gets copied
to the plugin path scalr.io/scalr/scalr. However, the correct location
(used by released versions and in the README) is
registry.scalr.io/scalr/scalr. This change updates the makefile to
install to a path that will be properly resolved in Terraform provider
configurations.