Match name exactly in data.scalr_vcs_provider
The name argument to the scalr_vcs_provider data source was being passed
directly to the list operation in the Scalr API, which uses full-text
search to filter its response.

This change keeps the same API logic, but adds an additional client-side
filter to match the name exactly, similar to the filters used by the
scalr_webhook and scalr_endpoint data sources.

It remains possible to omit the name argument as well (for example, if
an account only has one GitHub provider configuration).